Move sigmoid tests to respective files
Add tests for inplace Sigmoid and Softplus
